implement header as a TopBar component
